Experimenting with mat mul code generation

+++ apps/linear_algebra/benchmarks/macros.h
@@ -13,7 +13,7 @@
         }                                           \
     }
 
-#define L1GFLOPS(N) 2 * N * 1e-3 / elapsed
+#define L1GFLOPS(N) 2 * (float)N * 1e-3 / elapsed
 #define L1Benchmark(benchmark, type, code)                              \
     virtual void bench_##benchmark(int N) {                             \
         Scalar alpha = random_scalar();                                 \
@@ -31,7 +31,7 @@
                   << std::endl;                                         \
     }
 
-#define L2GFLOPS(N) (2 + N) * N * 1e-3 / elapsed
+#define L2GFLOPS(N) (float)(2 + N) * N * 1e-3 / elapsed
 #define L2Benchmark(benchmark, type, code)                              \
     virtual void bench_##benchmark(int N) {                             \
         Scalar alpha = random_scalar();                                 \
@@ -52,7 +52,7 @@
                   << std::endl;                                         \
     }
 
-#define L3GFLOPS(N) (3 + N) * N * N * 1e-3 / elapsed
+#define L3GFLOPS(N) (float)(3 + N) * N * N * 1e-3 / elapsed
 #define L3Benchmark(benchmark, type, code)                              \
     virtual void bench_##benchmark(int N) {                             \
         Scalar alpha = random_scalar();                                 \
